Remove default reference for framework
The default reference to the framework previously present in the project assumed the .framework lived in the parent directory of the project.

Instead advise the user via the README.md to drag in the .framework from their distribution as it may live in a different directory tree than this demo app.